+++ bot/main.py
@@ -29,7 +29,7 @@ from modules.database import init as db_init, get_all_items
 from modules.enumerations import PlayerStatus
 from modules.loader import init as cog_init, is_all_locked, unlock_all
 from modules.reactions import init as react_init, reaction_handler
-from modules.lobby import on_inactive_confirmed
+from modules.lobby import on_inactive_confirmed, lobby_init
 
 # Modules for the custom classes
 from modules.roles import init as roles_init, role_update, is_admin
@@ -292,6 +292,9 @@ def main(launch_str=""):
 
     # Initialise reaction handlers
     react_init(client)
+
+    # Init lobby
+    lobby_init(Match)
 
     # Add init handlers
     _add_init_handlers(client)